Fix redraw of SendTokenStackScreen
All components from it extracted to separate components/files.
